refactor: tighten bootstrap runtime boundary
